Revert "Origin/frontend/layout bug fixes"
