+++ packages/frontend/src/app/components/MultiChain/CurrentBlockchain/Mobile/CurrentBlockChainModile.tsx
@@ -5,8 +5,8 @@ import React, { useEffect } from 'react'
 import { useChangeNetwork } from '../../../../hooks/useChangeNetwork'
 import { useCurrentBlockChain } from '../../../../hooks/useCurrentBlockChain'
 import { useMultiChainStore } from '../../../../hooks/useMultiChainStore'
+import BoxShadowed from '../../../../pages/MainPage/components/BoxShadowed/BoxShadowed'
 import { Txt } from '../../../../UIkit'
-import BoxShadowed from '../../../../UIkit/BoxShadowed/BoxShadowed'
 import { ICurrentBlockchain } from '../../helper/types/currentBlockChainTypes'
 import { CurrentBlockchainStyle } from '../CurrentBlockchain.styles'
 import CurrentBlockchainBlock from '../CurrentBlockchainBlock/CurrentBlockchainBlock'
